requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
google-auth==2.38.0
google-api-python-client==2.187.0
PyMuPDF>=1.23.0
requests>=2.28.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pinecone>=3.0.0
apscheduler>=3.10.0
# openai>=1.0.0
google-genai>=1.0.0
